Extract shared node delete button and strengthen delete-control tests
