+++ tensormap-frontend/src/components/DragAndDropCanvas/CustomNodes/ConvNode/ConvNode.jsx
@@ -1,8 +1,8 @@
 import PropTypes from "prop-types";
-import { Handle, Position, useReactFlow } from "reactflow";
+import { Handle, Position } from "reactflow";
+import NodeDeleteButton from "../shared/NodeDeleteButton";
 
 function ConvNode({ data, id }) {
-  const { deleteElements } = useReactFlow();
   const p = data.params;
   const parts = [
     p.filter ? `F: ${p.filter}` : null,
@@ -14,27 +14,12 @@ function ConvNode({ data, id }) {
     .filter(Boolean)
     .join(", ");
 
-  const handleDelete = (event) => {
-    event.preventDefault();
-    event.stopPropagation();
-    deleteElements({ nodes: [{ id }] });
-  };
-
   return (
     <div className="w-48 rounded-lg border bg-white shadow-sm">
       <Handle type="target" position={Position.Left} isConnectable id={`${id}_in`} />
       <div className="flex items-center justify-between rounded-t-lg bg-node-conv px-3 py-1.5 text-xs font-bold text-white">
         <span>Conv2D</span>
-        <button
-          type="button"
-          onClick={handleDelete}
-          className="rounded px-1 leading-none text-white/80 hover:bg-white/20 hover:text-white"
-          aria-label="Delete layer"
-          title="Delete layer"
-          data-testid="conv-node-delete-button"
-        >
-          ×
-        </button>
+        <NodeDeleteButton id={id} testId="conv-node-delete-button" />
       </div>
       <div className="px-3 py-2 text-xs text-muted-foreground">{parts || "Not configured"}</div>
       <Handle type="source" position={Position.Right} isConnectable id={`${id}_out`} />
