+++ tensormap-frontend/src/components/DragAndDropCanvas/CustomNodes/ConvNode/ConvNode.test.jsx
@@ -101,8 +101,9 @@ describe("ConvNode", () => {
   it("renders delete button and deletes node on click", () => {
     render(<ConvNode {...defaultProps} />);
 
-    const deleteButton = screen.getByTestId("conv-node-delete-button");
+    const deleteButton = screen.getByRole("button", { name: "Delete layer" });
     expect(deleteButton).toBeInTheDocument();
+    expect(deleteButton).toHaveAttribute("title", "Delete layer");
 
     fireEvent.click(deleteButton);
     expect(deleteElementsMock).toHaveBeenCalledWith({ nodes: [{ id: defaultProps.id }] });
