+++ tensormap-frontend/src/components/DragAndDropCanvas/CustomNodes/DenseNode/DenseNode.test.jsx
@@ -78,8 +78,9 @@ describe("DenseNode", () => {
   it("renders delete button and deletes node on click", () => {
     render(<DenseNode {...defaultProps} />);
 
-    const deleteButton = screen.getByTestId("dense-node-delete-button");
+    const deleteButton = screen.getByRole("button", { name: "Delete layer" });
     expect(deleteButton).toBeInTheDocument();
+    expect(deleteButton).toHaveAttribute("title", "Delete layer");
 
     fireEvent.click(deleteButton);
     expect(deleteElementsMock).toHaveBeenCalledWith({ nodes: [{ id: defaultProps.id }] });
